Output coverage report and report only src/lib.rs
Add markdown output and `cat` it to print a plain text coverage report.
This helped figure out that the GitHub workflow was including a ton of
files from `/usr/share/cargo/registry` in the report, so tell `grcov` to
only keep `src/lib.rs`.

Teach `test-cover` to tell `grcov` about another directory to
search when `PGDATA` and `PGUSER` are both set. Required because
`--runas` does not copy `LLVM_PROFILE_FILE` to the new environment, so
the profile data files are instead written to the current directory in
which `--runas` tests are run. This happens to be "$PGDATA/$PGVERSION",
presumably the default for pgrx.

Also: box `CompileError` to make clippy happy. It is a large error, and
clippy complained about its size in more recent versions, causing the
lint test to fail.