Upgrade pgrx to v0.15.0
Requires changes to enum handling to avoid a race condition that v0.15.0
no recognizes and complains about. Done by adding code that tests the
Draft GUC indirectly. Also requires removing the `///` documentation
comments in the Draft enum, though that should be fixed in v0.15.1.
Otherwise requires changing string arguments to C strings.

pgrx v0.15.0 now passes environment variables starting with `LLVM_` to
the `--runas` user, but an attempt to set `LLVM_PROFILE_FILE` to write
profile files to `/tmp` causes the Postgres user to write no files at
all. Super annoying. So tweak `.ci/test-cover` to only set
`LLVM_PROFILE_FILE` if `PGUSER` is not set, and to look for files in
`PWD` and the data directory when it is.